poprawione testy productitemow - bigdecimal

+++ src/main/java/tech/allegro/wildsnake/showcase/model/ShowcaseItem.java
@@ -1,12 +1,14 @@
 package tech.allegro.wildsnake.showcase.model;
 
+import java.math.BigDecimal;
+
 public class ShowcaseItem {
     private final String title;
     private final String imageUrl;
-    private Long price;
+    private BigDecimal price;
 
 
-    public ShowcaseItem(String title, String imageUrl, Long price) {
+    public ShowcaseItem(String title, String imageUrl, BigDecimal price) {
         this.title = title;
         this.imageUrl = imageUrl;
         this.price = price;
@@ -20,7 +22,7 @@ public class ShowcaseItem {
         return imageUrl;
     }
 
-    public Long getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 }
